requirements.txt
websockets>=10.0
python-dotenv>=0.19.0
openai>=0.28.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
zhipuai>=2.0.1
requests>=2.28.0